Set default request rate to 1 per second for the boostr sync script (#240)
## Problem Statement

I am not sure that the current request rate setting will allow the
script to do a full sync given the Boostr API's rate limits, based on
some quick tests I've run. As we're preparing to merge our scheduled job
to run this script in all environments on a regular basis, I thought it
would be helpful to be able to configure this value.

## Proposed Changes

* Set up a parameter for the request rate instead of hard coding it,
meaning this value can be configurable without making a PR and pushing a
code change here.
* Set the default value of requests every 1 seconds for the script. At
this default setting, I have managed a successful full sync of the
current boostr data in 45 minutes.
* A few additional documentation updates, fixes, and helpful tips

## Verification Steps

The specs verify that the Boostr api wrapper code does in fact wait
between requests, and that a specified interval is used by the wrapper.
Let me know if you see any other opportunities for adding tests.

For manual verification, you can run the script locally (see
`syncBoostrData.md` for usage details) and pass a very high value to see
if the script is respecting the value you passed.