Gracefully handle 204 responses from /v1/tiles on Preview page (#271)
## Problem Statement

The /v1/tiles endpoint returns a 204 response with no body when there
are no tiles available. Currently when this happens, get_amp_tiles
raises an exception, and we don't display any ads (even if e.g. SPOCs
might be available):

![Screenshot 2024-10-03 at 2 50
12 PM](https://github.com/user-attachments/assets/f05c6690-c032-4dbf-896e-d47745e21048)

## Proposed Changes

This changes get_amp_tiles to gracefully handle a 204 response from
/v1/tiles and allows the preview page to display other ads that migth be
available.

## Verification Steps

This can be tricky to verify, as tiles _should_ almost always available
for the countries/regions currently supported by the preview page.
However, at time of writing, they are not available for Canada, as seen
in the above screenshot.

To verify:
1. Go to the Preview page
2. Pick a country/region where tiles are not available, and hit Preview
3. Observe that SPOCs are shown

## Visuals

![Screenshot 2024-10-03 at 2 50
27 PM](https://github.com/user-attachments/assets/a087a32e-85dd-4c2a-aada-513a87cbca80)