Revert "GHA maintenance (#274)" (#279)
This reverts commit 22bb3f76f8cef6d5c63c577935309a53e6cdd217.

## Problem Statement

I merged André's PR for updates to the github actions. I think I viewed
it in an earlier stage when it was just adding node linting and testing
to the PR. It looks like the final change added `lint` and `test`
workflows, but it looks like there is additional environment setup
needed for the test workflows to succeed. Here's an example of the
failed run:


https://github.com/mozilla-services/consvc-shepherd/actions/runs/11296454474/job/31421725378

Also, I think that if the `test` step fails, the `build and push` step
should not proceed.

## Proposed Changes

Revert the commit for now to make sure GHA pipelines are functioning as
expected. We can try this change again next week.

## Verification Steps

No verification needed, this is a revert commit so we'll bring back
previous GHA pipelines behavior.