dont search for flight/campaign name
This is a fix for a bug we were having when running the script again in the
same day. The problem was that when we ran it twice in the same day we would
get the error: `CommandError: duplicate key value violates unique constraint
"unique_delivered_flight"`.

This issue is happening because campaign_name and flight_name are being
included in the lookup condition. As a result, update_or_create is searching
for an exact match on campaign name, flight name, in addition to our unique
  constraints submission date, campaign id, flight id, and provider. When it
  doesn't find a matching record (such as when those non-nullable fields aren't
  an exact match), it attempts to create a new entry, but there's already a
  record with the same unique constraints combination, leading to the duplicate
  key error.

By moving the campaign name and flight name fields into defaults, they are only
updated or set and not used in the lookup.