[SHEP-36]: Adding make command to show human readable timestamps in local docker logs (#287)
## References

[SHEP-36](https://mozilla-hub.atlassian.net/browse/SHEP-36)

## Problem Statement

The original ask was to modify the timestamp in the log outputs to
display in a more human readable format to allow for easier local
debugging as by default, docker logs in the command line do not show
human readable timestamps. (They do in Docker Desktop however, but this
is not as convenient.)

## Proposed Changes

The [JSON formatting
library](https://python-dockerflow.readthedocs.io/en/main/api/logging.html)
we are using for logging conforms to the [MozLog
schema](https://wiki.mozilla.org/Firefox/Services/Logging). As a result,
we cannot (and probably shouldn't) modify the timestamp field to be in a
different format.

However, to achieve the same goal of making local dev logs more easy to
parse, we can make some changes to how we run our docker containers and
add a new make command `dev-ts` that will display timestamps next to all
docker logs. This however does come with a downside that we have to run
`docker-compose up` in detached mode, meaning the usual `ctrl-c` motion
will simply close out of the log follower, not kill the containers.

In addition, for scripts run through manage.py from inside the docker
container we've included the option for setting a local env variable
which will append human readable datetime headers in front of each log.
To do this the following changes were made:

1. Added a new (optional) env variable APP_ENV. This may be useful in
the future for other things as well and we may want to require it down
the line, but for now it can be omitted without issue.
2. Added a datetime header in front of all logs emitted when running
scripts via manage.py. These headers will only be included if APP_ENV ==
"dev" and previous behavior (with no datetime headers) is expected with
any other value for APP_ENV or if it is undefined.

Finally, we also updated the docs to add info about the schema of our
Json logs.


## Verification Steps
1.  **Verify console docker timestamps** 
- [ ] run `make dev-ts`
- [ ] Verify human-readable timestamps are visible next to logs.

2. **Verify default behavior in scripts is no datetime headers on logs**
- [ ] verify no `.env` variable exists named `APP_ENV`
- [ ] run `python manage.py sync_boostr_data
https://app.boostr.com/api/` inside the docker container
- [ ] verify logs look as they did before this PR (no datetime headers)

3. **Verify new behavior in scripts is _with_ datetime headers on logs**
- [ ] add a new `.env` variable with `APP_ENV=dev`
- [ ] run `python manage.py sync_boostr_data
https://app.boostr.com/api/` inside the docker container
- [ ] verify logs now have datetime headers in a newline above each log.


## Visuals
New behavior with `make dev-ts` command.
![Screenshot 2024-10-15 at 5 55
05 PM](https://github.com/user-attachments/assets/e6ebc735-5f2c-415e-ad90-4e5556e8230f)

New behavior in scripts with `APP_ENV=dev`
![Screenshot 2024-10-16 at 1 10
37 PM](https://github.com/user-attachments/assets/3148b5c2-33ca-46d3-8cdf-9e4f791f2880)


[SHEP-36]:
https://mozilla-hub.atlassian.net/browse/SHEP-36?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ