[SHEP-36] Hotfix for failing script in staging.
The error: django.core.exceptions.ImproperlyConfigured: Requested
setting USE_I18N, but settings are not configured. You must either
define the environment variable DJANGO_SETTINGS_MODULE or call
settings.configure() before accessing settings.

Is showing up in staging when trying to run the script. I cannot seem to
recreate this locally... but this sould be the issue based on the error
message